requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose
pydantic
PyPDF2
faiss-cpu
sentence-transformers
openai
requests
python-dotenv
